requirements.txt
cozmo==1.4.9
cozmoclad==3.3.0
six==1.12.0
websocket-client==0.50.0
